requirements.txt
faker
mysql-connector-python
